[🚀 사이클2 - 미션 (예약 대기)] 에버(조성진) 미션 제출합니다. (#476)
* docs: 구현 기능 추가

* docs: 구현 기능 추가

* feat: 예약 취소 시 대기자 예약 전환 기능 구현

* docs: 구현 기능 완료 표시

* refactor: 예약 취소 & 대기 취소 동시성 제어 로직 추가

* refactor: 예약 수정 시 대기자 예약 전환 로직 추가

* docs: 구현 기능 완료 체크

* refactor(ReservationService): 코드 리팩토링

- 도메인 책임 분리
- 에러 코드 수정
- 메서드 시그니처 변경

* refactor(ThemeService): 코드 리팩토링

- 에러 코드 static import

* refactor(TimeService): 코드 리팩토링

- 시간 삭제 로직 리팩토링

* refactor(ReservationWaitingService): 코드 리팩토링

- 에러 코드 static import
- 일시 검증 로직 리팩토링

* refactor: API URI 수정

* refactor: 프론트 코드 수정

* refactor: 서비스 클래스에 @Transactional 설정

* test: 예약 취소 & 수정 중 예외 발생 시 롤백 테스트 코드 작성

* refactor: waitingNumber가 없는 도메인 객체 생성

* refactor: 미사용 메서드 삭제

* refactor: 미사용 메서드 삭제

* refactor: 락 적용 테이블 단일화

* refactor: 락 적용 기간 단축

* test: 실패 테스트 코드 수정

+++ src/main/java/roomescape/time/service/TimeService.java
@@ -1,19 +1,21 @@
 package roomescape.time.service;
 
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.global.exception.ErrorCode;
 import roomescape.global.exception.RoomescapeException;
-import roomescape.reservation.Reservation;
 import roomescape.reservation.dao.ReservationDao;
 import roomescape.time.AvailableTime;
 import roomescape.time.ReservationTime;
 import roomescape.time.dao.TimeDao;
 
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.List;
+
+import static roomescape.global.exception.ErrorCode.CANNOT_DELETE_RESERVED_TIME;
+
 @Service
+@Transactional(readOnly = true)
 public class TimeService {
 
     private final ReservationDao reservationDao;
@@ -37,6 +39,7 @@ public class TimeService {
                 .toList();
     }
 
+    @Transactional
     public ReservationTime add(LocalTime startAt) {
         ReservationTime time = new ReservationTime(startAt);
         return timeDao.insert(time);
@@ -44,16 +47,9 @@ public class TimeService {
 
     @Transactional
     public void deleteById(Long id) {
-        List<Reservation> reservations = reservationDao.selectByTimeId(id);
-        for (Reservation reservation : reservations) {
-            validateExistTimeByReservation(reservation.getDate(), reservation.getTime(), id);
+        if (reservationDao.existsUpcomingByTimeId(id, LocalDate.now(), LocalTime.now())) {
+            throw new RoomescapeException(CANNOT_DELETE_RESERVED_TIME);
         }
         timeDao.deleteById(id);
-    }
-
-    private void validateExistTimeByReservation(LocalDate date, ReservationTime time, Long timeId) {
-        if (!time.isBeforeDateTime(date, time) && time.getId().equals(timeId)) {
-            throw new RoomescapeException(ErrorCode.CANNOT_DELETE_RESERVED_TIME);
-        }
     }
 }
